requirements.txt
st_pages==0.4.3
scikit-learn==1.3.0
plotly==5.15.0
pyparsing==3.1.1
streamlit-extras==0.3.0
